src/main/persistence/JsonWriter.java
package persistence;

import model.TimeDatabase;
import org.json.JSONObject;

import java.io.*;

// A writer that writes JSON representation of time database to file
// NOTE: much of this code is borrowed from the JsonWriter class of the WorkRoomApp example repo of CPSC 210
public class JsonWriter {
    private static final int TAB = 4;
    private PrintWriter writer;
    private String destination;

    // EFFECTS: constructs writer to write to destination file
    public JsonWriter(String destination) {
        this.destination = destination;
    }

    // MODIFIES: this
    // EFFECTS: opens writer; throws FileNotFoundException if destination file cannot
    // be opened for writing
    public void openTimeDatabaseFile() throws FileNotFoundException {
        writer = new PrintWriter(new File(destination));
    }

    // MODIFIES: this
    // EFFECTS: writes JSON representation of workroom to file
    public void write(TimeDatabase tdb) {
        JSONObject json = tdb.toJson();
        saveToFile(json.toString(TAB));
    }

    // MODIFIES: this
    // EFFECTS: closes writer
    public void close() {
        writer.close();
    }

    // MODIFIES: this
    // EFFECTS: writes string to file
    private void saveToFile(String json) {
        writer.print(json);
    }
}
